Fix pr comments.
