+++ flow-maven-plugin/src/main/java/com/vaadin/flow/plugin/maven/NodeUpdateAbstractMojo.java
@@ -32,7 +32,6 @@ import com.vaadin.flow.server.Command;
 import com.vaadin.flow.server.Constants;
 import com.vaadin.flow.server.frontend.ClassFinder;
 import com.vaadin.flow.server.frontend.FrontendUtils;
-import com.vaadin.flow.server.frontend.NodeUpdateImports;
 import org.apache.maven.artifact.DependencyResolutionRequiredException;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.logging.Log;
@@ -40,6 +39,8 @@ import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.project.MavenProject;
 import org.reflections.Reflections;
 import org.reflections.util.ConfigurationBuilder;
+
+import static com.vaadin.flow.server.frontend.FrontendUtils.FLOW_IMPORTS_FILE;
 
 /**
  * Common stuff for node update mojos.
@@ -126,7 +127,7 @@ public abstract class NodeUpdateAbstractMojo extends AbstractMojo {
     /**
      * A Flow JavaScript file with all project's imports to update.
      */
-    @Parameter(defaultValue = "${project.build.directory}/" + NodeUpdateImports.FLOW_IMPORTS_FILE)
+    @Parameter(defaultValue = "${project.build.directory}/" + FLOW_IMPORTS_FILE)
     protected File generatedFlowImports;
 
     /**
@@ -135,19 +136,19 @@ public abstract class NodeUpdateAbstractMojo extends AbstractMojo {
     @Parameter(defaultValue = "${project.basedir}/frontend")
     protected File frontendDirectory;
 
-    protected Command updater;
+    private Command updater;
 
     @Override
     public void execute() {
-        // Do nothing when bower mode
-        if (isBowerMode(getLog())) {
-            String goal = "update-frontend";
-            getLog().info("Skipped '" + goal + "' goal because `vaadin.bowerMode` is set.");
-            return;
-        }
 
         long start = System.nanoTime();
-        getUpdater().execute();
+
+        if (updater == null) {
+            updater = createUpdater();
+        }
+
+        updater.execute();
+
         long ms = (System.nanoTime() - start) / 1000;
         getLog().info("Took " + ms + "ms.");
     }
@@ -169,7 +170,7 @@ public abstract class NodeUpdateAbstractMojo extends AbstractMojo {
         return bowerMode;
     }
 
-    protected abstract Command getUpdater();
+    protected abstract Command createUpdater();
 
     static ClassFinder getClassFinder(MavenProject project) {
         final List<String> runtimeClasspathElements;
