+++ flow-server/src/main/java/com/vaadin/flow/server/frontend/FrontendUtils.java
@@ -41,6 +41,32 @@ public class FrontendUtils {
      */
     public static final String WEBPACK_CONFIG ="webpack.config.js";
 
+    /**
+     * NPM package name that will be used for the javascript files present in
+     * jar resources that will to be copied to the npm folder so as they are
+     * accessible to webpack.
+     */
+    public static final String FLOW_PACKAGE = "@vaadin/flow-frontend/";
+
+    /**
+     * File that contains Flow application imports, javascript, and theme annotations.
+     * It is also the entry-point for webpack.
+     */
+    public static final String FLOW_IMPORTS_FILE = "frontend/generated-flow-imports.js";
+    /**
+     *
+     * A parameter for overriding the
+     * {@link FrontendUtils#FLOW_IMPORTS_FILE} default value for the file
+     * with all Flow project imports.
+     */
+    public static final String MAIN_JS_PARAM = "vaadin.frontend.jsFile";
+
+    /**
+     * A special prefix to use in the webpack config to tell webpack to look for
+     * the import starting with a prefix in the Flow project frontend directory.
+     */
+    public static final String WEBPACK_PREFIX_ALIAS = "Frontend/";
+
     private static final String NOT_FOUND =
             "%n%n======================================================================================================"
             + "%nFailed to determine '%s' tool."
@@ -93,6 +119,17 @@ public class FrontendUtils {
     }
 
     /**
+     * Gets the flow package inside node_modules.
+     *
+     * @param nodeModulesPath
+     *            path to node_modules.
+     * @return the flow package folder.
+     */
+    public static File getFlowPackage(File nodeModulesPath) {
+        return new File(nodeModulesPath, FLOW_PACKAGE);
+    }
+
+    /**
      * Check that the folder structure does not meet a proper npm mode project.
      * It is useful to run V13 projects in V14 before they have been migrated.
      *
@@ -104,7 +141,6 @@ public class FrontendUtils {
         boolean hasNpmConfig = new File(getBaseDir(), PACKAGE_JSON).exists()
                 && new File(getBaseDir(), WEBPACK_CONFIG).exists();
         
-
         return hasBowerFrontend && !hasNpmFrontend && !hasNpmConfig ? true : false;
     }
